new: Connection to Docker PostgreSQL database, new Campaign layout, new controller methods

+++ archimedes/campaign-service/src/main/java/archimedes/campaignservice/CampaignServiceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CampaignServiceRepository extends JpaRepository<Campaign, Long> {
-	// TODO: ?
+public interface CampaignServiceRepository extends JpaRepository<Campaign, Integer> {
+	// Add specific queries here
 }
